refactor(zkevm): split the TrieNode guest variants into std/zkevm files
The single-threaded guest forms of the flag accessors and the RLP seqlock
were carried as `#if ZK_EVM` blocks inside TrieNode.cs, which put both
variants of four members in front of every reader of the file.

Move them to TrieNode.std.cs / TrieNode.zkevm.cs, the file-variant
convention Directory.Build.targets already applies to the rest of the
guest. TrieNode was partial, so this needs no new type surface, and the
`partial` here is the shared-base-plus-variant case rather than a marker
added to satisfy the file-name analyzer.

The std bodies are moved verbatim.
